+++ packages/index.js
@@ -4,14 +4,15 @@ import AlanButton from './Button'
 import CodeBackgroundWall from "./CodeBackgroundWall"
 import AlanBackTop from "./BackTop"
 import DynamicCard from "./DynamicCard"
+import TakingPictures from "./TakingPictures"
 import TherMometer from "./TherMometer"
 import YuanWar from "./YuanWar"
 import "./styles/index.scss"
 
 //按需引入
-export { StarrySky, DigitalScroll, AlanButton, CodeBackgroundWall, AlanBackTop, TherMometer , YuanWar, DynamicCard};
+export { StarrySky, DigitalScroll, AlanButton, CodeBackgroundWall, AlanBackTop, TherMometer , YuanWar, DynamicCard, TakingPictures};
 
-const components = [StarrySky, DigitalScroll, AlanButton, CodeBackgroundWall, AlanBackTop, TherMometer, YuanWar, DynamicCard];
+const components = [StarrySky, DigitalScroll, AlanButton, CodeBackgroundWall, AlanBackTop, TherMometer, YuanWar, DynamicCard, TakingPictures];
 
 const install = (App) => {
   components.forEach((item) => {